fix install
